aptos_service.py
"""
Aptos Blockchain Integration Service
Handles all interactions with Aptos smart contracts
"""

import os
from dotenv import load_dotenv
from typing import Optional, List, Dict, Any

load_dotenv()

# Mock Aptos service for development
# In production, use actual Aptos SDK
class AptosService:
    def __init__(self):
        # Initialize Aptos client
        self.node_url = os.getenv("APTOS_NODE_URL", "https://fullnode.testnet.aptoslabs.com/v1")
        self.contract_address = os.getenv("APTOS_CONTRACT_ADDRESS", "")
    
    def create_account(self) -> dict:
        """Create a new Aptos account (mock)"""
        import random
        import string
        address = '0x' + ''.join(random.choices(string.hexdigits.lower(), k=64))
        return {
            "address": address,
            "private_key": '0x' + ''.join(random.choices(string.hexdigits.lower(), k=64)),
            "public_key": '0x' + ''.join(random.choices(string.hexdigits.lower(), k=64))
        }
    
    def get_account_balance(self, address: str) -> int:
        """Get account balance in octas (mock)"""
        return 1000000000  # 10 APT
    
    def fund_account_from_faucet(self, address: str, amount: int = 100000000) -> bool:
        """Fund account from testnet faucet (mock)"""
        return True
    
    async def mint_dataset_nft(
        self, 
        owner_address: str,
        dataset_id: int,
        hash_str: str
    ) -> dict:
        """Mint a dataset NFT (mock)"""
        import hashlib
        tx_hash = '0x' + hashlib.sha256(f"{owner_address}{dataset_id}{hash_str}".encode()).hexdigest()
        return {
            "success": True,
            "transaction_hash": tx_hash,
            "dataset_id": dataset_id
        }
    
    async def grant_license(
        self,
        owner_address: str,
        dataset_id: int,
        user_address: str,
        duration_secs: int,
        license_type: str
    ) -> dict:
        """Grant a license to access a dataset (mock)"""
        import hashlib
        tx_hash = '0x' + hashlib.sha256(f"{owner_address}{dataset_id}{user_address}".encode()).hexdigest()
        return {
            "success": True,
            "transaction_hash": tx_hash
        }
    
    async def set_dataset_price(
        self,
        owner_address: str,
        dataset_id: int,
        base_price: float,
        per_query_price: float
    ) -> dict:
        """Set pricing for a dataset (mock)"""
        import hashlib
        tx_hash = '0x' + hashlib.sha256(f"{owner_address}{dataset_id}{base_price}".encode()).hexdigest()
        return {
            "success": True,
            "transaction_hash": tx_hash
        }
    
    async def pay_for_license(
        self,
        buyer_address: str,
        seller_address: str,
        dataset_id: int,
        amount: float
    ) -> dict:
        """Pay for a dataset license (mock)"""
        import hashlib
        tx_hash = '0x' + hashlib.sha256(f"{buyer_address}{seller_address}{dataset_id}".encode()).hexdigest()
        return {
            "success": True,
            "transaction_hash": tx_hash
        }
    
    async def set_royalty_config(
        self,
        owner_address: str,
        dataset_id: int,
        contributors: List[str],
        share_percentage: int
    ) -> dict:
        """Set royalty configuration for a dataset (mock)"""
        import hashlib
        tx_hash = '0x' + hashlib.sha256(f"{owner_address}{dataset_id}".encode()).hexdigest()
        return {
            "success": True,
            "transaction_hash": tx_hash
        }
    
    def get_transaction_status(self, tx_hash: str) -> dict:
        """Get transaction status (mock)"""
        return {
            "success": True,
            "vm_status": "Executed successfully",
            "hash": tx_hash
        }

# Singleton instance
aptos_service = AptosService()
