requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
aptos-sdk==0.11.0
attrs==25.4.0
behave==1.3.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cucumber-expressions==18.0.1
cucumber-tag-expressions==8.1.0
ecdsa==0.19.1
fastapi==0.122.0
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
multidict==6.7.0
parse==1.20.2
parse_type==0.6.6
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyNaCl==1.6.1
python-dotenv==1.2.1
python-graphql-client==0.4.3
python-multipart==0.0.20
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.44
starlette==0.50.0
tomli==2.3.0
types-six==1.17.0.20251009
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
yarl==1.22.0
